tests/test_api.py
"""
Unit tests for API endpoints.
"""
import pytest
from fastapi.testclient import TestClient
from PIL import Image
import io

# Note: Import will fail if models aren't loaded, so we mock for testing
# from src.api.main import app


@pytest.fixture
def client():
    """Create test client."""
    # This is a placeholder - actual implementation would need mocked models
    from fastapi import FastAPI
    app = FastAPI()
    
    @app.get("/health")
    def health():
        return {"status": "healthy", "models_loaded": True, "version": "1.0.0"}
    
    return TestClient(app)


def test_health_endpoint(client):
    """Test health check endpoint."""
    response = client.get("/health")
    assert response.status_code == 200
    data = response.json()
    assert "status" in data
    assert "models_loaded" in data
    assert "version" in data


def create_test_image():
    """Create a test image."""
    img = Image.new('RGB', (224, 224), color='red')
    img_byte_arr = io.BytesIO()
    img.save(img_byte_arr, format='PNG')
    img_byte_arr.seek(0)
    return img_byte_arr


# Additional tests would require mocking the models
# Example structure:

# def test_predict_endpoint(client):
#     """Test prediction endpoint."""
#     img = create_test_image()
#     response = client.post(
#         "/predict",
#         files={"image": ("test.png", img, "image/png")},
#         data={"review_text": "Great product!"}
#     )
#     assert response.status_code == 200
#     data = response.json()
#     assert "category" in data
#     assert "sentiment" in data
#     assert "recommendation_score" in data


if __name__ == "__main__":
    pytest.main([__file__, "-v"])