requirements.txt
# Core Dependencies
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
tokenizers==0.14.1


# Data Processing
numpy==1.24.3
pandas==2.1.0
Pillow==10.0.1
opencv-python==4.8.1.78
albumentations==1.3.1
scikit-learn==1.3.0

# NLP
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2

# API & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
streamlit==1.28.0

# Visualization & Monitoring
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
tensorboard==2.15.0
wandb==0.16.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
requests==2.31.0

# Explainability
grad-cam==1.4.8
captum==0.6.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.1

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4